Solved bug with Bomb, Solved bug with player latency, made a normal MainMenu

+++ Assets/Scripts/Runtime/MonoBehaviours/UI/GameplayPresenter.cs
@@ -51,7 +51,7 @@ namespace Runtime.MonoBehaviours.UI
         }
 
         private void UpdateHealthText(int newValue) => HealtText.text = $"{newValue}";
-        private void UpdateSpeedText(int newValue) => SpeedText.text = $"{newValue}";
+        private void UpdateSpeedText(float newValue) => SpeedText.text = $"{newValue}";
         private void UpdateBombsDamageText(int newValue) => BombsDamageText.text = $"{newValue}";
         private void UpdateSpreadText(int newValue) => SpreadText.text = $"{newValue}";
         private void UpdateBombsPerTimeText(int newValue) => BombsPerTimeText.text = $"{newValue}";
